refactor: Centralize governance type definitions, make `Proposal.createdAt` mandatory, and add Hyperliquid chain to execution details.
