+++ packages/summer-earn-gov-validator/src/components/ProposalExecutionDetails.tsx
@@ -3,9 +3,10 @@
 import React, { useMemo, useState } from 'react'
 import { useAccount, useSwitchChain, useWriteContract } from 'wagmi'
 
+import { CrossChainProposal,Proposal } from '@/types/governance'
+
 import { ChainTheme, getChainTheme } from '../config/chains'
 import config from '../config/index.json'
-import { CrossChainProposal, Proposal } from '../services/subgraph'
 import {
   addresToContractName,
   decodeCalldata,
@@ -37,6 +38,7 @@ const CHAIN_ID_TO_NETWORK: Record<string, keyof typeof config> = {
   '8453': 'base',
   '42161': 'arbitrum',
   '146': 'sonic',
+  '999': 'hyperliquid',
 }
 
 interface ProposalExecutionDetailsProps {
